Restructuration of routing and adding active link
