+++ content/ClientApp/app.js
@@ -1,6 +1,6 @@
 ﻿import Vue from 'vue'
 import axios from 'axios'
-import router from './router'
+import router from './router/index'
 import store from './store'
 import { sync } from 'vuex-router-sync'
 import App from 'components/app-root'
